[ENH] Strip ``run_command`` out of ``interfaces.base``
